+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   between `fn` and `(`.
 - Fixed a bug where the formatter would incorrectly remove blocks around some
   binary operators.
+- Constants can now be defined after they are used in functions
 
 ## v0.12.1 - 2020-11-15